+++ quasar.conf.js
@@ -10,7 +10,6 @@ module.exports = function (ctx) {
       'mxmConfig',
       'apollo',
       'mxmVal',
-      'mxmProviderClient',
       'utl'
     ],